cbmc-library/Float* cleanup: rename or move to cbmc/
In cbmc-library/, we organise regression tests by the name of the
library function they are exercising. Rename these tests with the
library function they exercise. For those not focussing on library
functions, move them to cbmc/ instead. As those tests are also run past
the cprover SMT solver, three tests newly demonstrated issues in the SMT
back-end.
