--- regression/cbmc-library/Float-to-double1/test.desc
@@ -1,8 +1,0 @@
-CORE
-main.c
---floatbv
-^EXIT=0$
-^SIGNAL=0$
-^VERIFICATION SUCCESSFUL$
---
-^warning: ignoring
