+++ regression/cbmc-library/isnan-01/main.c
@@ -1,9 +1,19 @@
 #include <assert.h>
 #include <math.h>
 
+#if defined(_WIN32) && !defined(__CYGWIN__)
+#include <float.h>
+#define isnan _isnan
+#endif
+
+float nondet_float();
+
 int main()
 {
-  isnan();
-  assert(0);
-  return 0;
+  float f = nondet_float();
+
+  double d = (double)f;
+  float ff = (float)d;
+
+  assert((f == ff) || (isnan(f) && isnan(ff)));
 }
