+++ regression/cbmc/Float-flags-no-simp1/test.desc
@@ -1,4 +1,4 @@
-CORE
+CORE broken-smt-backend thorough-paths
 main.c
 --floatbv --no-simplify
 ^EXIT=0$
