+++ regression/cbmc/Float-no-simp9/test.desc
@@ -1,4 +1,4 @@
-CORE
+CORE broken-smt-backend thorough-paths
 main.c
 --floatbv --no-simplify
 ^EXIT=0$
